feat: test all api service
