+++ backend/user/requirements.txt
@@ -2,6 +2,7 @@ django==5.1.5
 djangorestframework
 djangorestframework-simplejwt
 djangorestframework-queryfields
+requests
 pillow
 drf-spectacular==0.25.1
 psycopg2
