+++ backend/user/server/config/settings.py
@@ -136,5 +136,8 @@ MEDIA_ROOT = os.path.join(BASE_DIR, MEDIA_URL)
 REST_FRAMEWORK = {
     'DEFAULT_AUTHENTICATION_CLASSES': (
         'rest_framework_simplejwt.authentication.JWTAuthentication',
+    ),
+    'DEFAULT_PARSER_CLASSES': (
+        'rest_framework.parsers.JSONParser',
     )
 }
